requirements.txt
sqlalchemy==2.0.40
fastapi==0.115.12
psycopg2-binary==2.9.10
pydantic==2.11.3
uvicorn==0.34.2
requests==2.32.3
python-dotenv==1.1.0